Move prepare into separate module
