+++ src/pages.js
@@ -5,8 +5,8 @@ function outputPages (pages, context, options, handlebars) {
   // write me
 }
 
-function buildPages (options, context, handlebars) {
-  outputPages(context.pages, context, options, handlebars);
+function buildPages (drizzleData) {
+  // outputPages(context.pages, context, options, handlebars);
 }
 
 export default buildPages;
